requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
idna==2.10
multidict==4.7.6
PyYAML==5.3.1
telepot==12.7
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.5.1
